Stream dir/path renamed to output log path
