+++ crates/hyperqueue/src/stream/reader/mod.rs
@@ -1,1 +1,1 @@
-pub mod streamdir;
+pub mod outputlog;
